+++ R/utils.R
@@ -321,13 +321,11 @@ look <- function(x) {
 #' Print More Rows of Ichimoku Objects
 #'
 #' After calling or invoking the default print method for ichimoku objects, the
-#'     console output will display \code{# … with x more rows} if the entire
-#'     data does not fit on-screen. Use \code{more()} to display more rows.
+#'     console output will display \code{-- omitted x rows} if the entire data
+#'     does not fit on-screen. Use \code{more()} to display more rows.
 #'
-#' @param n a parameter (optional) passed on to the argument 'n' of
-#'     \code{\link[tibble]{print.tbl}} controlling the number of rows to show.
-#'     Defaults to 100. Normally an integer, but supply a character string such
-#'     as 'a' to print all rows.
+#' @param rows (optional) specify the number of rows to print; defaults to all
+#'     rows if not supplied or non-numeric.
 #'
 #' @return The ichimoku object contained in \code{\link{.Last.value}} (invisibly)
 #'     or else invisible NULL (if .Last.value is not an ichimoku object).
@@ -336,15 +334,15 @@ look <- function(x) {
 #' @examples
 #' cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
 #' cloud
+#' more(25)
 #' more()
-#' more(20)
 #'
 #' @export
 #'
-more <- function(n) {
+more <- function(rows) {
 
   is.ichimoku(lv <- .Last.value) || return(invisible())
-  print(lv, plot = FALSE, n = if (missing(n)) 100 else n)
+  print(lv, plot = FALSE, rows = if (missing(rows) || !is.numeric(rows)) attr(lv, "dim")[1L] else rows)
 
 }